build: move unconvert into megacheck
This also checks tests, which were previously not covered.

+++ pkg/sql/distsqlrun/disk_row_container_test.go
@@ -100,7 +100,7 @@ func TestDiskRowContainer(t *testing.T) {
 		},
 	}
 
-	rng := rand.New(rand.NewSource(int64(timeutil.Now().UnixNano())))
+	rng := rand.New(rand.NewSource(timeutil.Now().UnixNano()))
 
 	evalCtx := parser.MakeTestingEvalContext()
 	diskMonitor := mon.MakeMonitor(
